Move ForgeLocalMod to API, document
